+++ CHANGELOG.md
@@ -3738,3 +3738,4 @@
 - [2026-08-19 10:45] test(i18n): add unit tests for i18n business logic
 - [2026-08-19 11:08] test(network): add unit tests for network business logic
 - [2026-08-19 11:37] docs(auth): update readme and API contracts for auth
+- [2026-08-19 12:10] test(auth): implement widget integration tests for auth